Move declarations and definitions to header file
